New API - web platform code comments doc ready
This is just cleanup of comments and reorg of the directory
so the source code is ready for documentation generation.
