+++ src/platform/web/index.ts
@@ -1,8 +1,4 @@
-import * as Modifiers from "./modifiers";
-export { Modifiers };
-export { SessionDescriptionHandler } from "./session-description-handler";
-export { SimpleUser } from "./simple-user";
-export { SimpleUserDelegate } from "./simple-user-delegate";
-export { SimpleUserOptions } from "./simple-user-options";
-export { Transport } from "./transport";
-export { TransportOptions } from "./transport-options";
+export * from "./modifiers";
+export * from "./session-description-handler";
+export * from "./simple-user";
+export * from "./transport";
