+++ src/platform/web/simple-user/index.ts
@@ -1,0 +1,7 @@
+/**
+ * A simple SIP user implementation for web browsers.
+ * @packageDocumentation
+ */
+export * from "./simple-user";
+export * from "./simple-user-delegate";
+export * from "./simple-user-options";
