--- src/platform/web/transport-options.ts
@@ -1,10 +1,0 @@
-export interface TransportOptions {
-  /** URLs of one or more WebSocket servers to connect with. */
-  wsServers: Array<string> | string;
-  connectionTimeout?: number;
-  maxReconnectionAttempts?: number;
-  reconnectionTimeout?: number;
-  keepAliveInterval?: number;
-  keepAliveDebounce?: number;
-  traceSip?: boolean;
-}
